refactor: useSingletonModal
