+++ src/modals/ComposeModal.tsx
@@ -2,6 +2,7 @@
 
 import { Dialog, Transition } from '@headlessui/react';
 import { XMarkIcon } from '@heroicons/react/24/outline';
+import { useSingletonModal } from '@masknet/shared-base-ui';
 import { forwardRef, Fragment, useCallback, useState } from 'react';
 
 import LoadingIcon from '@/assets/loading.svg';
@@ -10,7 +11,6 @@ import Discard from '@/components/Compose/Discard.js';
 import WithLexicalContextWrapper from '@/components/Compose/WithLexicalContextWrapper.js';
 import { EMPTY_LIST } from '@/constants/index.js';
 import type { SingletonModalRefCreator } from '@/maskbook/packages/shared-base/src/index.js';
-import { useSingletonModal } from '@/maskbook/packages/shared-base-ui/src/index.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 import type { IPFS_MediaObject } from '@/types/index.js';
