+++ src/modals/LensStatusModal.tsx
@@ -2,6 +2,7 @@
 
 import { Dialog, Transition } from '@headlessui/react';
 import { Trans } from '@lingui/macro';
+import { useSingletonModal } from '@masknet/shared-base-ui';
 import { forwardRef, Fragment } from 'react';
 import { useDisconnect } from 'wagmi';
 
@@ -11,7 +12,6 @@ import { PlatformIcon } from '@/components/PlatformIcon.js';
 import { SocialPlatform } from '@/constants/enum.js';
 import { Image } from '@/esm/Image.js';
 import type { SingletonModalRefCreator } from '@/maskbook/packages/shared-base/src/index.js';
-import { useSingletonModal } from '@/maskbook/packages/shared-base-ui/src/index.js';
 import { LoginModalRef, LogoutModalRef } from '@/modals/controls.js';
 import { useLensStateStore } from '@/store/useLensStore.js';
